fix overrides and images on nodejs
