+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/). Every release, along with the migration instructions, is documented on the Github [Releases](https://github.com/airbnb/react-sketchapp/releases) page.
 
+## Version 3.0.0-beta.2
+
+- Fix setting overrides (#409)
+- Fix images on NodeJS
+
 ## Version 3.0.0-beta.1
 
 - Fix ShapeGroup on nodejs (#387)
