--- __tests__/jest/jsonUtils/hacksForJSONImpl.js
@@ -1,15 +1,0 @@
-import * as hacks from '../../../src/jsonUtils/hacksForJSONImpl';
-
-describe('API', () => {
-  it('exports makeImageDataFromUrl', () => {
-    expect(hacks.makeImageDataFromUrl).toBeInstanceOf(Function);
-  });
-
-  it('exports makeImageDataFromUrl', () => {
-    expect(hacks.createAttributedString).toBeInstanceOf(Function);
-  });
-
-  it('exports makeEncodedAttributedString', () => {
-    expect(hacks.makeEncodedAttributedString).toBeInstanceOf(Function);
-  });
-});
