+++ package.json
@@ -44,6 +44,7 @@
     "airbnb-prop-types": "^2.10.0",
     "error-stack-parser": "^2.0.0",
     "invariant": "^2.2.2",
+    "js-sha1": "^0.6.0",
     "murmur2js": "^1.0.0",
     "normalize-css-color": "^1.0.1",
     "pegjs": "^0.10.0",
@@ -118,7 +119,7 @@
     "react": "^16.3.2",
     "react-test-renderer": "^16.3.2",
     "rimraf": "^2.5.4",
-    "sketchapp-json-flow-types": "^0.3.4"
+    "sketchapp-json-flow-types": "^0.3.5"
   },
   "skpm": {
     "test": {
